+++ Changelog.rst
@@ -16,6 +16,9 @@ version 3.17.0
 * Fix bug that caused `Data._axes` to be incorrect after a call to
   `cf.Field.collapse`
   (https://github.com/NCAS-CMS/cf-python/issues/857)
+* Fix bug that caused wrong directions from
+  `cf.DimensionCoordinate.direction`
+  (https://github.com/NCAS-CMS/cf-python/issues/859)
 * Changed dependency: ``Python>=3.9.0``
 * Changed dependency: ``numpy>=2.0.0``
 * Changed dependency: ``cfdm>=1.12.0.0, <1.12.1.0``